# LATCH_GO_PORT_API_SPEC.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Latch API",
    "version": "1.0.0",
    "description": "Machine-readable spec for porting Latch Next.js /api/* routes to Go. Extension-verified. See LATCH_GO_PORT_GUIDE.md."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Go dev server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Next.js reference server"
    }
  ],
...
  "paths": {
...
    "/api/transaction/build-delegated": {
      "post": {
        "summary": "POST /api/transaction/build-delegated",
        "operationId": "_api_transaction_build-delegated_post",
        "tags": [
          "transaction"
        ],
        "x-latch-source-file": "app/api/transaction/build-delegated/route.ts",
        "x-latch-auth": "none",
        "x-latch-env-required": [
          "BUNDLER_SECRET",
          "NEXT_PUBLIC_FACTORY_ADDRESS",
          "NEXT_PUBLIC_VERIFIER_ADDRESS",
          "NEXT_PUBLIC_WEBAUTHN_VERIFIER_ADDRESS",
          "NEXT_PUBLIC_COUNTER_ADDRESS",
          "NEXT_PUBLIC_RPC_URL",
          "NEXT_PUBLIC_NETWORK_PASSPHRASE"
        ],
        "x-latch-db-models": [],
        "x-latch-legacy-error": true,
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "txXdr": {
                      "type": "string"
                    },
                    "smartAccountAuthEntryXdr": {
                      "type": "string"
                    },
                    "gAddressPreimageXdr": {
                      "type": "string"
                    },
                    "gAddressEntryTemplateXdr": {
                      "type": "string"
                    },
                    "authDigestHex": {
                      "type": "string"
                    },
                    "validUntilLedger": {
                      "type": "integer"
                    },
                    "contextRuleId": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "txXdr": {
                      "type": "string"
                    },
                    "smartAccountAuthEntryXdr": {
                      "type": "string"
                    },
                    "gAddressPreimageXdr": {
                      "type": "string"
                    },
                    "gAddressEntryTemplateXdr": {
                      "type": "string"
                    },
                    "authDigestHex": {
                      "type": "string"
                    },
                    "validUntilLedger": {
                      "type": "integer"
                    },
                    "contextRuleId": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "smartAccountAddress",
                  "gAddress"
                ],
                "properties": {
                  "smartAccountAddress": {
                    "type": "string"
                  },
                  "gAddress": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sid",
        "description": "30-day sliding session. Auto-created on first request."
      }
    },
    "schemas": {
...
      "LegacyErrorBody": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}